feat(refactor): improve readability, accessibility and performance
